extracted isValidWiki to new file
