--- src/components/CreateWiki/CreateWikiToast.tsx
@@ -1,1 +1,0 @@
-const CreateWikiToast = () => {}
