+++ src/pages/create-wiki/index.tsx
@@ -63,14 +63,12 @@ import {
 import { logEvent } from '@/utils/googleAnalytics'
 import {
   initialMsg,
-  MINIMUM_WORDS,
   useCreateWikiState,
   CreateWikiProvider,
   useGetSignedHash,
   useCreateWikiEffects,
   useCreateWikiContext,
   defaultErrorMessage,
-  isVerifiedContentLinks,
   isWikiExists,
   ValidationErrorMessage,
   sanitizeContentToPublish,
@@ -85,10 +83,9 @@ import {
 import useConfetti from '@/hooks/useConfetti'
 import WikiScoreIndicator from '@/components/Editor/WikiScoreIndicator'
 import { useWhiteListValidator } from '@/hooks/useWhiteListValidator'
-import { MEDIA_POST_DEFAULT_ID, WIKI_SUMMARY_LIMIT } from '@/data/Constants'
 import CreateWikiPageHeader from '@/components/SEO/CreateWikiPage'
-import { getWordCount } from '@/utils/DataTransform/getWordCount'
 import { getWikiMetadataById } from '@/utils/WikiUtils/getWikiFields'
+import { isValidWiki } from '@/utils/CreateWikiUtils/isValidWiki'
 
 type PageWithoutFooter = NextPage & {
   noFooter?: boolean
@@ -174,84 +171,8 @@ const CreateWikiContent = () => {
     return slug
   }
 
-  const isValidWiki = () => {
-    if (!wiki.title) {
-      toast({
-        title: `Add a Title at the top for this Wiki to continue `,
-        status: 'error',
-        duration: 3000,
-      })
-      return false
-    }
-
-    if (wiki.title.length > 60) {
-      toast({
-        title: `Title should be less than 60 characters`,
-        status: 'error',
-        duration: 3000,
-      })
-      return false
-    }
-
-    const words = getWordCount(wiki.content || '')
-
-    if (words < MINIMUM_WORDS) {
-      toast({
-        title: `Add a minimum of ${MINIMUM_WORDS} words in the content section to continue, you have written ${words}`,
-        status: 'error',
-        duration: 3000,
-      })
-      return false
-    }
-
-    if (!isVerifiedContentLinks(wiki.content)) {
-      toast({
-        title: 'Please remove all external links from the content',
-        status: 'error',
-        duration: 3000,
-      })
-      return false
-    }
-
-    if (!wiki.images?.length) {
-      toast({
-        title: 'Add a main image on the right column to continue',
-        status: 'error',
-        duration: 3000,
-      })
-      return false
-    }
-
-    if (wiki.categories.length === 0) {
-      toast({
-        title: 'Add one category to continue',
-        status: 'error',
-        duration: 3000,
-      })
-      return false
-    }
-
-    if (!wiki.media?.every(m => !m.id.endsWith(MEDIA_POST_DEFAULT_ID))) {
-      toast({
-        title: 'Some of media are still uploading, please wait',
-        status: 'error',
-        duration: 3000,
-      })
-      return false
-    }
-
-    if (wiki.summary && wiki.summary.length > WIKI_SUMMARY_LIMIT) {
-      toast({
-        title: `Summary exceeds maximum limit of ${WIKI_SUMMARY_LIMIT}`,
-        status: 'error',
-        duration: 3000,
-      })
-    }
-    return true
-  }
-
   const saveOnIpfs = async (override?: boolean) => {
-    if (!isValidWiki()) return
+    if (!isValidWiki(toast, wiki)) return
 
     logEvent({
       action: 'SUBMIT_WIKI',
